Moving Average

+++ src/main/kotlin/us/albertim/fan/db/FanControlTable.kt
@@ -4,6 +4,6 @@ import org.jetbrains.exposed.dao.id.IntIdTable
 
 object FanControlTable : IntIdTable() {
   val timestamp = long("timestamp").defaultExpression(CurrentTimestampExpression())
-  val percent = integer("percent")
+  val percent = double("percent")
   val auto = bool("auto")
 }
